requirements-dev.txt
# Development dependencies
-r requirements.txt

python-dotenv>=1.0.0


